requirements.txt
aiohttp==3.6.2
amqp==2.6.0
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-filter==2.2.0
djangorestframework==3.11.0
future==0.18.2
idna==2.9
kombu==4.6.11
Markdown==3.2.1
multidict==4.7.5
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
slackclient==2.5.0
sqlparse==0.3.0
urllib3==1.25.9
vine==1.3.0
wget==3.2
yarl==1.4.2
